+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.0.2
 * Make `ResponseContext` *explicitly* implement `StreamConsumer` (though technically it already did???)
+* Split `Controller.configureServer` to create `Controller.applyRoutes`.
 
 # 2.0.1
 * Tracked down a bug in `Driver.runPipeline` that allowed fallback